# api/gateway/analysis/models/data_model_list.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"strconv"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// DataModelList data model list
//
// swagger:model DataModelList
type DataModelList struct {

	// data models
	// Required: true
	DataModels []*DataModelSummary `json:"dataModels"`

	// paging
	// Required: true
	Paging *Paging `json:"paging"`
}

// Validate validates this data model list
func (m *DataModelList) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateDataModels(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePaging(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *DataModelList) validateDataModels(formats strfmt.Registry) error {

	if err := validate.Required("dataModels", "body", m.DataModels); err != nil {
		return err
	}

	for i := 0; i < len(m.DataModels); i++ {
		if swag.IsZero(m.DataModels[i]) { // not required
			continue
		}

		if m.DataModels[i] != nil {
			if err := m.DataModels[i].Validate(formats); err != nil {
				if ve, ok := err.(*errors.Validation); ok {
					return ve.ValidateName("dataModels" + "." + strconv.Itoa(i))
				}
				return err
			}
		}

	}

	return nil
}

func (m *DataModelList) validatePaging(formats strfmt.Registry) error {

	if err := validate.Required("paging", "body", m.Paging); err != nil {
		return err
	}

	if m.Paging != nil {
		if err := m.Paging.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("paging")
			}
			return err
		}
	}

	return nil
}

// MarshalBinary interface implementation
func (m *DataModelList) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *DataModelList) UnmarshalBinary(b []byte) error {
	var res DataModelList
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}